extract screens into their own files
